Move unique connections to its own namespace
while it currently comes from the radio-threshold service, and it feelslike a threshold for a radio to pas for rewarding; current radio thresholds is about service provider boosting, and unique connections is about oracle boosting.
