+++ mobile_verifier/src/lib.rs
@@ -17,6 +17,7 @@ pub mod speedtests_average;
 pub mod subscriber_location;
 pub mod subscriber_verified_mapping_event;
 pub mod telemetry;
+pub mod unique_connections;
 
 pub use settings::Settings;
