+++ mobile_verifier/src/unique_connections/mod.rs
@@ -1,0 +1,7 @@
+pub mod db;
+pub mod ingestor;
+
+use helium_crypto::PublicKeyBinary;
+use std::collections::HashMap;
+
+pub type UniqueConnectionCounts = HashMap<PublicKeyBinary, u64>;
